requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.3
distro==1.9.0
Django==5.2.6
django-allauth==65.12.1
django-countries==7.6.1
django-extensions==4.1
django-formtools==2.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
oauthlib==3.3.1
openai==2.1.0
pillow==11.3.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
